backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
paddleocr==2.7.0.3
paddlepaddle==2.5.1
groq==0.4.2
Pillow==10.0.1
opencv-python==4.8.1.78
reportlab==4.0.7
